+++ app/frontend/src/features/calendarSelection/styles/CalendarStyle.ts
@@ -1,0 +1,143 @@
+import { StyleSheet } from "react-native";
+
+export const calendarStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#f8f8f8",
+  },
+  centerContainer: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    padding: 32,
+  },
+  header: {
+    backgroundColor: "white",
+    padding: 20,
+    borderBottomWidth: 1,
+    borderBottomColor: "#eee",
+  },
+  headerTitle: {
+    fontSize: 22,
+    fontWeight: "bold",
+    color: "#333",
+  },
+  headerSubtitle: {
+    fontSize: 14,
+    color: "#666",
+    marginTop: 6,
+  },
+  list: {
+    flex: 1,
+  },
+  listContent: {
+    padding: 16,
+    paddingBottom: 24,
+  },
+  calendarItem: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    backgroundColor: "white",
+    padding: 16,
+    borderRadius: 12,
+    marginBottom: 10,
+    elevation: 2,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 1 },
+    shadowOpacity: 0.1,
+    shadowRadius: 2,
+  },
+  calendarItemSelected: {
+    borderWidth: 2,
+    borderColor: "#912338",
+    backgroundColor: "#fff5f5",
+  },
+  calendarItemLeft: {
+    flexDirection: "row",
+    alignItems: "center",
+    flex: 1,
+  },
+  colorDot: {
+    width: 12,
+    height: 12,
+    borderRadius: 6,
+    marginRight: 12,
+  },
+  calendarName: {
+    fontSize: 16,
+    fontWeight: "600",
+    color: "#333",
+  },
+  primaryBadge: {
+    fontSize: 11,
+    color: "#912338",
+    marginTop: 2,
+  },
+  footer: {
+    backgroundColor: "white",
+    padding: 20,
+    paddingBottom: 36,
+    borderTopWidth: 1,
+    borderTopColor: "#eee",
+  },
+  selectedCount: {
+    fontSize: 14,
+    color: "#666",
+    marginBottom: 12,
+  },
+  confirmButton: {
+    backgroundColor: "#912338",
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: 8,
+    paddingVertical: 14,
+    borderRadius: 10,
+  },
+  confirmButtonDisabled: {
+    backgroundColor: "#ccc",
+    opacity: 0.8,
+  },
+  confirmButtonText: {
+    color: "white",
+    fontWeight: "bold",
+    fontSize: 16,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+    color: "#333",
+    textAlign: "center",
+  },
+  subtitle: {
+    fontSize: 14,
+    color: "#666",
+    textAlign: "center",
+    marginTop: 12,
+    paddingHorizontal: 24,
+  },
+  hint: {
+    fontSize: 12,
+    color: "#999",
+    textAlign: "center",
+    marginTop: 16,
+    fontStyle: "italic",
+  },
+  loadingText: {
+    fontSize: 14,
+    color: "#666",
+    marginTop: 12,
+  },
+  retryButton: {
+    marginTop: 20,
+    backgroundColor: "#912338",
+    paddingVertical: 12,
+    paddingHorizontal: 24,
+    borderRadius: 8,
+  },
+  retryButtonText: {
+    color: "white",
+    fontWeight: "bold",
+  },
+});
