+++ app/frontend/src/services/calendarListService.ts
@@ -1,31 +1,85 @@
 import axios from "axios";
 
-export interface GoogleCalendar {
+const GOOGLE_CALENDAR_API = "https://www.googleapis.com/calendar/v3";
+
+export interface GoogleCalendarListItem {
   id: string;
   summary: string;
   primary?: boolean;
+  backgroundColor?: string;
+  foregroundColor?: string;
+  accessRole?: string;
 }
 
-const GOOGLE_CALENDAR_API_URL =
-  "https://www.googleapis.com/calendar/v3/users/me/calendarList";
+export interface FetchCalendarListResult {
+  error: string | null;
+  message?: string;
+  calendars: GoogleCalendarListItem[];
+}
 
-export const fetchGoogleCalendars = async (
+/**
+ * Fetches the user's calendar list from Google Calendar API.
+ * Requires OAuth2 access token with calendar.readonly scope (from US-3.1).
+ */
+export const fetchCalendarList = async (
   accessToken: string
-): Promise<GoogleCalendar[]> => {
+): Promise<FetchCalendarListResult> => {
+  const url = `${GOOGLE_CALENDAR_API}/users/me/calendarList`;
+
   try {
-    const response = await axios.get(GOOGLE_CALENDAR_API_URL, {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
+    const response = await axios.get(url, {
+      headers: { Authorization: `Bearer ${accessToken}` },
+      params: { maxResults: 50 },
+      timeout: 10000,
     });
 
-    return response.data.items.map((calendar: any) => ({
-      id: calendar.id,
-      summary: calendar.summary,
-      primary: calendar.primary,
-    }));
-  } catch (error) {
-    console.error("Error fetching Google calendars:", error);
-    throw error;
+    const items = response.data.items ?? [];
+    const calendars: GoogleCalendarListItem[] = items.map(
+      (item: {
+        id: string;
+        summary?: string;
+        primary?: boolean;
+        backgroundColor?: string;
+        foregroundColor?: string;
+        accessRole?: string;
+      }) => ({
+        id: item.id,
+        summary: item.summary ?? "Untitled Calendar",
+        primary: item.primary ?? false,
+        backgroundColor: item.backgroundColor,
+        foregroundColor: item.foregroundColor,
+        accessRole: item.accessRole,
+      })
+    );
+
+    return { error: null, calendars };
+  } catch (error: unknown) {
+    if (axios.isAxiosError(error)) {
+      const status = error.response?.status;
+      const message = error.response?.data?.error?.message ?? error.message;
+
+      if (status === 401 || status === 403) {
+        return {
+          calendars: [],
+          error: "token_expired",
+          message:
+            status === 401
+              ? "Session expired. Please reconnect your Google Calendar."
+              : "Access denied. Please reconnect your Google Calendar.",
+        };
+      }
+
+      return {
+        calendars: [],
+        error: "api_error",
+        message: message || "Could not load calendars.",
+      };
+    }
+
+    return {
+      calendars: [],
+      error: "network_error",
+      message: "Network error. Check your connection.",
+    };
   }
 };
